refactor: put parseSafe functions in csv.service
